moved to layout for blog post
