+++ src/app/blog/[slug]/loading.tsx
@@ -1,22 +1,13 @@
 export default function LoadingSingleBlogPage() {
   return (
     <div className="animate-pulse">
-      <div className="mb-8">
-        <h2 className="h-7 w-[30ch] bg-zinc-300" />
-        <div className="mt-2 h-5 w-[12ch] bg-zinc-300" />
-      </div>
-
-      <div>
-        <div>
-          <div className="h-4 w-[60ch] bg-zinc-300" />
-          <hr className="my-5" />
-          <div className="space-y-4">
-            <div className="h-40 w-[60ch] bg-zinc-300" />
-            <div className="h-24 w-[60ch] bg-zinc-300" />
-            <div className="h-28 w-[60ch] bg-zinc-300" />
-            <div className="h-12 w-[60ch] bg-zinc-300" />
-          </div>
-        </div>
+      <div className="h-4 w-[60ch] bg-zinc-300 mt-4" />
+      <hr className="my-5" />
+      <div className="space-y-4">
+        <div className="h-40 w-[60ch] bg-zinc-300" />
+        <div className="h-24 w-[60ch] bg-zinc-300" />
+        <div className="h-28 w-[60ch] bg-zinc-300" />
+        <div className="h-12 w-[60ch] bg-zinc-300" />
       </div>
     </div>
   )
